Move unit tests into new unit test project.
